backend/pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.2",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "markitdown[all,pdf]>=0.1.2",
    "numpy>=2.3.4",
    "openai>=2.7.2",
    "pytz>=2025.2",
    "tiktoken>=0.12.0",
    "python-multipart",
    "uvicorn",
    "docling",
    "docling[ocr]",
    "torch",
    "torchvision",
    "easyocr",
    "rapidocr_onnxruntime",
    "onnxruntime",
    "websockets",
    "markdown",
    "reportlab"
]